Refactor: Banish Nexus quaternions from most places.
This is the first stage in getting rid of the Nexus library entirely. It's redundant with System.Numerics.Vectors, which is required by SharpGLTF on top of being the official implementation of these concepts (and has intrinsics in .net core, not that that matters here).

Remaining is type converters for the inspector and the nexus <=> system.numerics converters, the latter basically because of the former.