Add the port-rigging modelscript command
This copies all the bone transforms and bone-skin inverse binding
matrices from a Diesel model file into the currently-loaded model data.

This is useful for making custom models work with their original
animations.

It also has an optional (enabled by default) setting to move all the
models to the correct parent, to avoid fiddling with the rootpoint
setting. This is particularly useful if we ever get imports/exports of
colliders working, as they're not all placed on the same nodes.