gltf: Remap bone IDs for overwritten objeccts
When importing an object, GLTF has already assigned a mapping of bones
to their indexes, and used that to build the per-vertex weight data.

When importing a model, the SkinBones is not overwritten (to permit the
reuse of all the rigging, for compatibility with existing animations).
This means the SkinBones has it's own ideas of what ID corresponds to
what bone.

Previously there was not any conversion between these two values,
leading to broken rigging in-game. This patch fixes that, going over all
the models and re-mapping the IDs to match what the SkinBones expects.

This logic already existed for newly-created SkinBones, so this patch
moves it out into a separate function used by all skinned models.